+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Core: Missing link button in new dialogs.
 - Billing: Fixed problem where amounts got multiplied by 100 when editing numbers with decimals
 - Workflow: Fixed workflow delete button
+- Files: Fixed file access to address book access denied
 
 17-09-2019 6.4.56
 - Core: Disable logging during upgrade for performance